Change SubId from long to int type.

+++ src/java/com/android/internal/telephony/SubscriptionInfoUpdater.java
@@ -323,7 +323,7 @@ public class SubscriptionInfoUpdater extends Handler {
                 number.put(SubscriptionManager.NUMBER, msisdn);
                 contentResolver.update(SubscriptionManager.CONTENT_URI, number,
                         SubscriptionManager.UNIQUE_KEY_SUBSCRIPTION_ID + "="
-                        + Long.toString(subId), null);
+                        + Integer.toString(subId), null);
             }
 
             SubscriptionInfo subInfo = mSubscriptionManager.getActiveSubscriptionInfo(subId);
@@ -348,7 +348,7 @@ public class SubscriptionInfoUpdater extends Handler {
                     mContext.getString(com.android.internal.R.string.unknownName));
             contentResolver.update(SubscriptionManager.CONTENT_URI, name,
                     SubscriptionManager.UNIQUE_KEY_SUBSCRIPTION_ID
-                    + "=" + Long.toString(subId), null);
+                    + "=" + Integer.toString(subId), null);
             logd("carrier name = " + simCarrierName);
 
             /* Update preferred network type and network selection mode on SIM change.
